requirements.txt
pandas==2.1.4
requests==2.32.3
ultralytics==8.2.0
opencv-python==4.9.0.80
pillow==10.1.0
shapely==2.0.2
scikit-learn==1.4.2
openpyxl==3.1.2
numpy==1.26.0
torch==2.1.0
torchvision==0.16.0
